refactor(packages): removed package application. It was moved packages and classes in packages usecase and validators to package model.
