+++ src/main/java/br/com/fiapbook/user/model/validator/UserEmailAlreadyRegisteredValidator.java
@@ -1,4 +1,4 @@
-package br.com.fiapbook.user.application.validator;
+package br.com.fiapbook.user.model.validator;
 
 import br.com.fiapbook.shared.exception.DuplicatedException;
 import br.com.fiapbook.user.model.service.UserService;
